animal-live-stock/requirements.txt

Werkzeug==3.1.5






itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
blinker==1.9.0
click==8.3.1


Flask==3.1.2
flask-cors==4.0.0

tensorflow==2.16.2
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.6.0.66
h5py==3.11.0

pymongo==4.8.0
joblib==1.4.2